Change to channel based system
The old way of keeping track of order using a list was actually not
concurrency safe under heavy load, so this was changed to make use of
channels instead.
This does mean that New() now has to spawn a goroutine to keep track of
the new internal queue, but that's good enough.